+++ install-connector.ps1
@@ -3,7 +3,7 @@ param(
     [string]$Secret,
     [ValidateSet("amd64", "arm64")]
     [string]$Architecture = "",
-    [string]$Release = "2.50.3",
+    [string]$Release = "2.50.4",
     [switch]$Uninstall
 )